+++ pyccel/ast/builtin_methods/list_methods.py
@@ -188,12 +188,12 @@ class ListExtend(ListMethod):
     Represents a call to the .extend() method of an object with a list type,
     which adds items of an iterable (list, tuple, dictionary, etc) at the end
     of a list.
-    This method is handled through the call to `_visit_ListExtend` in
+    This method is handled through the call to `_build_ListExtend` in
     the semantic stage. It then attempts to construct a `For` loop node with
     a body that calls `append()`, or direct `append()` nodes depending on
     the type of the iterable passed to `extend()`.
     This class should never be instantiated; it's only purpose is to help
-    construct the annotation_method `_visit_ListExtend`. 
+    construct the annotation_method `_build_ListExtend`. 
     The extend method is called as follows:
 
     >>> a = [1, 2, 3]